Combined Assistant and Vector_Storage Classes into one file.
